+++ src/main/java/com/projectx/foundit/repository/ITokenRepository.java
@@ -14,7 +14,7 @@ public interface ITokenRepository  extends CrudRepository<Token, Integer> {
             select t from Token t inner join User u on t.user.id = u.id
             where u.id = :userId and (t.expired = false or t.revoked = false)
             """)
-    List<Token> findAllValidTokensByUser(Long userId);
+    List<Token> findAllValidTokensByUser(int userId);
 
     Optional<Token> findByToken(String token);
 }
